# openapi.yaml
openapi: 3.0.3
info:
  title: USBI Platform API
  description: |
    API institucional del sistema USBI. Todos los errores usan RFC 7807 (Problem Details).
    Todos los endpoints autenticados requieren JWT Bearer.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.usbi.edu.mx/api/v1
    description: Servidor de producción (On-Premise, TLS 1.2+)
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Registro de nuevo usuario
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '201':
          description: Usuario creado. Pendiente de consentimiento de tutor si es menor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      summary: Inicio de sesión
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Login exitoso. Devuelve JWT firmado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: Cierre de sesión (invalida JWT via token_version)
      operationId: logoutUser
      responses:
        '204':
          description: Sesión cerrada. token_version incrementado en DB.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      summary: Renovar sesión con refresh token opaco
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequestDTO'
      responses:
        '200':
          description: Tokens rotados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/tutor-consent:
    post:
      summary: Solicitud de consentimiento de tutor (paso 1 del doble opt-in)
      operationId: submitTutorConsent
      security: []
      description: |
        Inicia el doble opt-in del tutor. Guarda los datos del tutor como una
        solicitud PENDIENTE con un token de un solo uso (24h por defecto) y envía
        un enlace de verificación al correo del tutor. La cuenta del menor NO se
        activa aquí: la activación ocurre sólo cuando el tutor abre el enlace
        (GET /auth/tutor-consent/verify), momento en que se registran la IP, la
        fecha y la hora del clic como evidencia legal del consentimiento.

        La respuesta es intencionalmente uniforme (202) exista o no la cuenta y
        esté o no pendiente, para no filtrar la existencia de cuentas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorConsentRequestDTO'
      responses:
        '202':
          description: |
            Solicitud aceptada. Si la cuenta requiere consentimiento, se envió un
            correo de verificación al tutor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending_verification]
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Error de validación de los datos del tutor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '502':
          description: No se pudo enviar el correo de verificación al tutor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/tutor-consent/verify:
    get:
      summary: Verificación del consentimiento de tutor (paso 2 del doble opt-in)
      operationId: verifyTutorConsent
      security: []
      description: |
        Endpoint destinado al navegador del tutor (enlace del correo). Valida el
        token, registra el clic (IP, fecha y hora) como evidencia legal, crea el
        registro de consentimiento con sello HMAC y activa la cuenta del menor.
        Responde con una página HTML (no JSON) porque el tutor la abre desde su
        cliente de correo.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Token opaco de un solo uso entregado en el correo del tutor.
      responses:
        '200':
          description: Consentimiento confirmado; cuenta activada. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Token ausente o inválido. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '409':
          description: El token ya había sido utilizado. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '410':
          description: El token caducó. Página HTML.
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Error interno al procesar la verificación. Página HTML.
          content:
            text/html:
              schema:
                type: string
  /auth/age-up:
    post:
      summary: Transición manual a mayoría de edad (Ley 251 — Self-Service)
      operationId: ageUpTransition
      description: |
        Endpoint manual de auto-servicio. El usuario solicita explícitamente
        la actualización de is_adult=true. La transición minimiza los datos
        identificables del tutor asociados a la cuenta. Limitado a 3 intentos
        por cuenta, rastreados en DB. NO se ejecuta como cron job.
      responses:
        '200':
          description: Transición aceptada. is_adult actualizado.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Límite de 3 intentos excedido. Cuenta bloqueada para esta operación.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sync:
    post:
      summary: Sincronización de progreso offline (idempotente)
      operationId: syncOfflineProgress
      description: |
        Recibe progreso generado offline. Valida HMAC sobre una cadena técnica
        canónica derivada de sync_event_id, user_id, device_id, crypto_key_version
        y payload. El device_id debe haber sido registrado por el backend.
        Recalcula XP en servidor y aplica merge idempotente por sync_event_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncEventRequestDTO'
      responses:
        '200':
          description: Sincronización exitosa o ya procesada (idempotente).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncEventResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            No autenticado o JWT expirado/invalidado, **o** firma HMAC
            inválida sobre el payload de sync (posible intento de trampa).
            Ambos casos comparten este código HTTP; se distinguen por el
            campo `type` de `ProblemDetails`
            (`.../errors/unauthorized` vs. `.../errors/invalid-signature`).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: |
            Payload de sync inválido tras la verificación de HMAC (por
            ejemplo: fecha de intento futura, `device_id` no registrado o
            nivel no disponible).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /arco:
    post:
      summary: Solicitud de Derechos ARCO (Acceso, Rectificación, Cancelación, Oposición)
      operationId: submitArcoRequest
      description: |
        Registra una solicitud ARCO. Si el tipo es 'cancelacion', el backend
        marcará wipe_local_data=true en la tabla devices para el usuario,
        que será inyectado en la siguiente respuesta de sync.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArcoRequestDTO'
      responses:
        '201':
          description: Solicitud ARCO registrada con evidence_hash.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArcoResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /arco/pending:
    get:
      summary: Listar solicitudes ARCO pendientes
      operationId: listPendingArcoRequests
      description: Sólo admin/director. Devuelve metadatos técnicos sin PII.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Solicitudes pendientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArcoPendingListDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /arco/{request_id}/resolve:
    post:
      summary: Resolver solicitud ARCO
      operationId: resolveArcoRequest
      description: |
        Sólo admin/director. Si la solicitud aprobada es de cancelación,
        pseudonimiza PII, revoca sesiones y marca dispositivos para borrado local.
      parameters:
        - in: path
          name: request_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveArcoRequestDTO'
      responses:
        '200':
          description: Solicitud resuelta.
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/security-incidents:
    post:
      summary: Registrar un incidente de seguridad (bitácora del Documento de Seguridad)
      operationId: createSecurityIncident
      description: |
        Sólo admin/director. Registra un incidente en la bitácora de retención de
        5 años, sellado con un hash HMAC de evidencia (No-Repudio) y con entrada en
        admin_audit_log. Los campos de seguimiento (reported_at, resolved_at,
        notified_*) se completan posteriormente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityIncidentRequestDTO'
      responses:
        '201':
          description: Incidente registrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  severity:
                    type: string
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Error de validación del incidente.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /devices:
    get:
      summary: Listar dispositivos registrados del usuario autenticado
      operationId: listDevices
      responses:
        '200':
          description: Dispositivos activos del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponseDTO'
    post:
      summary: Registrar dispositivo para sincronización offline
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequestDTO'
      responses:
        '201':
          description: Dispositivo registrado. Usar `id` como device_id en sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDTO'
  /sections:
    get:
      summary: Listar secciones oficiales
      operationId: listSections
      parameters:
        - in: query
          name: include_unpublished
          schema:
            type: boolean
            default: false
          description: Sólo admin/operator/director puede ver borradores.
      responses:
        '200':
          description: Lista de secciones visibles para el rol autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionListResponseDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Crear sección oficial
      operationId: createSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequestDTO'
      responses:
        '201':
          description: Sección creada como borrador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sections/{section_id}:
    patch:
      summary: Actualizar sección oficial
      operationId: updateSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequestDTO'
      responses:
        '200':
          description: Sección actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
        '404':
          $ref: '#/components/responses/NotFound'
  /sections/{section_id}/publish:
    post:
      summary: Publicar sección
      operationId: publishSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección publicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
  /sections/{section_id}/unpublish:
    post:
      summary: Despublicar sección
      operationId: unpublishSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección regresada a borrador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
  /sections/{section_id}/archive:
    post:
      summary: Archivar sección
      operationId: archiveSection
      parameters:
        - $ref: '#/components/parameters/SectionID'
      responses:
        '200':
          description: Sección archivada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionDTO'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT firmado con HS256. Incluye user_id, role y token_version.
  schemas:
...
    RegisterRequestDTO:
      type: object
      required: [full_name, email, password, is_adult, privacy_notice_version]
      additionalProperties: false
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        is_adult:
          type: boolean
        privacy_notice_version:
          type: string
    RegisterResponseDTO:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, pending_tutor_consent]
        message:
          type: string
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Conflict:
      description: Conflicto de estado (ej. UUID duplicado, email ya registrado)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...